Fix gcc 4.7 compilation problem for power_manager (again).
TEST=built using gcc 4.7
BUG=None

Reviewed-on: https://gerrit.chromium.org/gerrit/27828
Tested-by: Han Shen <shenhan@chromium.org>
Commit-Ready: Han Shen <shenhan@chromium.org>

+++ power_manager/input.cc
@@ -536,7 +536,8 @@ gboolean Input::EventHandler(GIOChannel* source, GIOCondition condition,
     LOG(ERROR) << "Short read in Input::EventHandler";
     return true;
   }
-  if ((num_events * sizeof(struct input_event)) != read_size) {
+  if ((num_events * sizeof(struct input_event)) !=
+      static_cast<size_t>(read_size)) {
     LOG(WARNING) << "Read size, not a discrete number of input_events in "
                  << "Input::EventHandler";
   }
